Generatore E-commerce/requirements.txt
Flask==2.0.2
waitress
openpyxl
openai
azure-storage-blob
waitress
